Design files for the resume builder
Here is the UI for the resume builder. Annoyingly, it will contain everything we already have from the keyword matcher, as I could not pull just that page, but i deleted the redundant HTML files.
Other than that: We need to find a nice image for the upload cloud. I don't have one with a transparent backgournd.
On form submit, we should have it so that the user can enter their email address, which triggers us sending our finished files to them. What the form submit looks like, is available if you see the success css, its basically just a box at this point.